Blog post from our Season of KDE 2022 student Karanjot Singh
This is a blog post written by our Season of KDE 2022 student Karanjot Singh. I have proofread the post and am satisfied with the text and images for publication on the KDE Eco blog.

Note: We had some issues with Git when pushing to the repository, and I hope we have done things correctly to successfully merge. If not, my apologies in advance. I am still getting used to using Git for larger projects.